src/backend/requirements.txt
grpcio
pyyaml
psycopg2
boto3
sqlitedict
torchvision
